requirements.txt
streamlit
streamlit-drawable-canvas
numpy
Pillow